+++ datasets/v1/collections.go
@@ -1,4 +1,4 @@
-package datasets // import "github.com/tilebox/tilebox-go/datasets/v1"
+package datasets
 
 import (
 	"context"